rag_system/requirements.txt
colpali-engine
PyMuPDF
Pillow
transformers==4.51.0
torch==2.4.1
torchvision==0.19.1
lancedb
rank_bm25
fuzzywuzzy
python-Levenshtein